// sandbox/src/app/docs/maps/new/providers/google/enum/e-adress-type.ts
export enum AddressType {
    country = 'country',
    locality = 'locality',
    postal_town = 'postal_town',
    administrative_area_level_1 = 'administrative_area_level_1',
    administrative_area_level_2 = 'administrative_area_level_2',
    route = 'route',
    street_number = 'street_number',
    postal_code = 'postal_code',
}